@@ -39,7 +39,7 @@ import { DagreComponent } from 'share/common/dagre/dagre.component';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class JobOverviewComponent implements OnInit, OnDestroy {
-  @ViewChild(DagreComponent) dagreComponent: DagreComponent;
+  @ViewChild(DagreComponent, { static: true }) dagreComponent: DagreComponent;
   nodes: NodesItemCorrectInterface[] = [];
   links: NodesItemLinkInterface[] = [];
   destroy$ = new Subject();
